frontend/requirements.txt
# Frontend
streamlit==1.29.0
PyYAML==6.0.1
plotly==5.20.0